% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{spatialTools-package}
\alias{spatialTools}
\alias{spatialTools-package}
\title{Spatial Toolbox}
\description{
\tabular{ll}{
Package: \tab tmap\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2015-04-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
Miscellaneous spatial tools, such as creating smooth polylines from points.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS}

